requirements (1).txt
python-dotenv
gradio
requests
playwright==1.54.0
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
langchain==0.1.20
langchain-community==0.0.38
langchain-core==0.1.52
langchain-experimental==0.0.52
langchain-google-genai==0.0.6
typing-extensions
pydantic==1.10.12
google-search-results
langgraph==0.0.20
beautifulsoup4
wikipedia
reportlab==4.2.2
